challenge-02-my-rainbow-sorcerer/src/app/requirements.txt
PyCryptodome
